prepare for cran update

+++ src/gplm0.cpp
@@ -69,7 +69,7 @@ Rcpp::List gplm0_density_evaluation_unknown_c_cpp(const arma::vec& theta,
     arma::vec yp = arma::vec(X * x).subvec(0, n-1);
     arma::vec ypo = yp + arma::randn(n) % arma::sqrt(varr);
     double log_lik = 0.0;
-    for(size_t i = 0; i < n; ++i) {
+    for(int i = 0; i < n; ++i) {
         log_lik += log_of_normal_pdf(y(i), yp(i), std::sqrt(varr(i)));
     }
     return Rcpp::List::create(
@@ -137,7 +137,7 @@ Rcpp::List gplm0_density_evaluation_known_c_cpp(const arma::vec& theta,
     arma::vec yp = arma::vec(X * x).subvec(0, n-1);
     arma::vec ypo = yp + arma::randn(n) % arma::sqrt(varr);
     double log_lik = 0.0;
-    for(size_t i = 0; i < n; ++i) {
+    for(int i = 0; i < n; ++i) {
         log_lik += log_of_normal_pdf(y(i), yp(i), std::sqrt(varr(i)));
     }
     return Rcpp::List::create(
